Update css to fix ribbon scrollbar overlap
Previously the 'Fork me on GitHub' image overlapped the scrollbar.

This was most notable on Windows, but also existed to some extent on Mac
OS and Linux distros.

Problem: The visible scrollbar did not belong to the body element, and
so the full width of the window was not taken into account by the
position:fixed style of the ribbon.

Solution: The solution was to ensure that only the scrollbar from the
body element is used and visible.

Visually checked on Chrome/Firefox/Safar on Mac OS, Chrome/IE11 on
Windows and Firefox on Debian.